feat(wgpu): add I32/U32 kernels across matmul, reduce, sort, and scatter
WebGPU's integer coverage stopped at a handful of shaders; most ops
had only an F32 pipeline. Adds I32/U32 WGSL shaders and launchers for
matmul (plain and fused-bias), reduce (full, arg, and int-accumulator
variants), sort and search (bitonic sort, searchsorted, top-k), scatter
(wide-accumulator scatter_reduce), clamp, linspace/arange, fused
elementwise (tensor and scalar forms), and the min/max-plus semiring
matmuls. `int_from_float.wgsl` and `int_wide_div.wgsl` hold the
shared int-domain building blocks the generated shaders reuse, and
`int_matmul_acc.wgsl` documents the 32-bit-limb accumulator convention
matmul and semiring matmul share, since WGSL has no 64-bit integer type.

+++ src/runtime/wgpu/ops/native/fused_elementwise.rs
@@ -7,7 +7,7 @@ use crate::runtime::wgpu::shaders::fused_elementwise;
 use crate::runtime::wgpu::{WgpuClient, WgpuRuntime};
 use crate::tensor::Tensor;
 
-/// Native fused_mul_add: out = a * b + c. F32 only.
+/// Native fused_mul_add: out = a * b + c. F32, I32, U32.
 pub(crate) fn native_fused_mul_add(
     client: &WgpuClient,
     a: &Tensor<WgpuRuntime>,
@@ -61,7 +61,7 @@ pub(crate) fn native_fused_mul_add(
     Ok(out)
 }
 
-/// Native fused_add_mul: out = (a + b) * c. F32 only.
+/// Native fused_add_mul: out = (a + b) * c. F32, I32, U32.
 pub(crate) fn native_fused_add_mul(
     client: &WgpuClient,
     a: &Tensor<WgpuRuntime>,
@@ -115,7 +115,7 @@ pub(crate) fn native_fused_add_mul(
     Ok(out)
 }
 
-/// Native fused_mul_add_scalar: out = a * scale + bias. F32 only.
+/// Native fused_mul_add_scalar: out = a * scale + bias. F32, I32, U32.
 pub(crate) fn native_fused_mul_add_scalar(
     client: &WgpuClient,
     a: &Tensor<WgpuRuntime>,
@@ -137,8 +137,8 @@ pub(crate) fn native_fused_mul_add_scalar(
         &out_buf,
         numel,
         dtype,
-        scale as f32,
-        bias as f32,
+        scale,
+        bias,
     )?;
 
     Ok(out)
